+++ docs/bookmarklets/opena11y-example.js
@@ -719,8 +719,8 @@
   }
 
   /*
-  *   normalize: Trim leading and trailing whitespace and condense all
-  *   internal sequences of whitespace to a single space. Adapted from
+  *   normalize: Trim leading and trailing white space and condense all
+  *   internal sequences of white space to a single space. Adapted from
   *   Mozilla documentation on String.prototype.trim polyfill. Handles
   *   BOM and NBSP characters.
   */
@@ -15787,7 +15787,7 @@
   *   @function isAriaHiddenFalse
   *
   *   @desc Returns true if the node has the aria-hidden property set to
-  *         "false", otherwise false.
+  *         "true", otherwise false.
   *         NOTE: This function is important in the accessible name
   *               calculation, since content hidden with a CSS technique
   *               can be included in the accessible name calculation when
@@ -15798,7 +15798,7 @@
   *   @return  see @desc
   */
 
-  function isAriaHiddenFalse(node) {
+  function isAriaHiddenTrue(node) {
 
     if (!node) {
       return false;
@@ -15810,7 +15810,7 @@
 
     if (node.nodeType === Node.ELEMENT_NODE) {
       return (node.hasAttribute('aria-hidden') &&
-          (node.getAttribute('aria-hidden').toLowerCase() === 'false'));
+          (node.getAttribute('aria-hidden').toLowerCase() === 'true'));
     }
 
     return false;
@@ -15833,18 +15833,14 @@
   function includeContentInName(node) {
     // NOTE: Chrome is the only major browser using aria-hidden=false in
     //       accessible name computation
-    const flag = isAriaHiddenFalse(node) && false;
-    return flag || (!isVisibilityHidden(node) && !isDisplayNone(node));
+    const flag = isAriaHiddenTrue(node);
+    return !flag && (!isVisibilityHidden(node) && !isDisplayNone(node));
   }
 
   /*
-  *   @function includeContentInName
+  *   @function getNodeContents
   *
-  *   @desc Checks the CSS display and hidden properties, and
-  *         the aria-hidden property to see if the content
-  *         should be included in the accessible name
-  *        calculation.  Returns true if it should be
-  *         included, otherwise false
+  *   @desc Get text content from a node for a name or description
   *
   *   @param  {Object}   node     -  DOM node
   *   @param  {Object}   forElem  -  DOM node the name is being computed for
@@ -15900,7 +15896,7 @@
                 contents = getEmbeddedControlValue(node);
               }
               else {
-                if (node.hasChildNodes()) {
+                if (!isAriaHiddenTrue(node) && node.hasChildNodes()) {
                   let children = Array.from(node.childNodes);
                   children.forEach( child => {
                     [nc, nInclAlt, nInclAriaLabel] = getNodeContents(child, forElem);
@@ -16489,6 +16485,10 @@
       this.ariaInfo  = new AriaInfo(accNameDoc, this.hasRole, this.role, defaultRole, elementNode, ariaVersion);
       this.eventInfo = new EventInfo(elementNode);
 
+      this.isInert   = elementNode.hasAttribute('inert') ?
+                       elementNode.inert :
+                       parentInfo.isInert;
+
       this.tabIndex             = checkTabIndex(elementNode);
       this.isTabStop            = checkIsTabStop(elementNode);
       this.isInteractiveElement = checkForInteractiveElement(elementNode);
@@ -16810,7 +16810,7 @@
    *
    * @desc Returns true if the element is natively interactive
    *
-   * @param  {Object}  node - DOM node
+   * @param  {Object}   node    - DOM node
    *
    * @return Returns true if the elements is interactive, otherwise false
    */
@@ -29079,6 +29079,7 @@
       this.useParentDocForName = false;
       this.documentIndex   = 0;
       this.domElement      = null;
+      this.isInert         = false;
       this.landmarkElement = null;
       this.listElement     = null;
       this.mapElement      = null;
@@ -29098,6 +29099,7 @@
         this.useParentDocForName = info.useParentDocForName;
         this.documentIndex   = info.documentIndex;
         this.domElement      = info.domElement;
+        this.isInert         = info.isInert;
         this.landmarkElement = info.landmarkElement;
         this.listElement     = info.listElement;
         this.mapElement      = info.mapElement;
@@ -29378,6 +29380,7 @@
       let newParentInfo = new ParentInfo(parentInfo);
       newParentInfo.domElement = domElement;
 
+      newParentInfo.isInert         = domElement.isInert;
       newParentInfo.controlElement  = this.controlInfo.update(controlElement, domElement);
       newParentInfo.mapElement      = this.imageInfo.update(mapElement, domElement);
       newParentInfo.inLink          = this.linkInfo.update(domElement, parentInfo.inLink);