+++ src/accName/namefrom.js
@@ -423,7 +423,7 @@ function isVisibilityHidden(node, psuedo=null) {
 *   @function isAriaHiddenFalse
 *
 *   @desc Returns true if the node has the aria-hidden property set to
-*         "false", otherwise false.
+*         "true", otherwise false.
 *         NOTE: This function is important in the accessible name
 *               calculation, since content hidden with a CSS technique
 *               can be included in the accessible name calculation when
@@ -434,7 +434,7 @@ function isVisibilityHidden(node, psuedo=null) {
 *   @return  see @desc
 */
 
-function isAriaHiddenFalse(node) {
+function isAriaHiddenTrue(node) {
 
   if (!node) {
     return false;
@@ -446,7 +446,7 @@ function isAriaHiddenFalse(node) {
 
   if (node.nodeType === Node.ELEMENT_NODE) {
     return (node.hasAttribute('aria-hidden') &&
-        (node.getAttribute('aria-hidden').toLowerCase() === 'false'));
+        (node.getAttribute('aria-hidden').toLowerCase() === 'true'));
   }
 
   return false;
@@ -469,18 +469,14 @@ function isAriaHiddenFalse(node) {
 function includeContentInName(node) {
   // NOTE: Chrome is the only major browser using aria-hidden=false in
   //       accessible name computation
-  const flag = isAriaHiddenFalse(node) && false;
-  return flag || (!isVisibilityHidden(node) && !isDisplayNone(node));
+  const flag = isAriaHiddenTrue(node);
+  return !flag && (!isVisibilityHidden(node) && !isDisplayNone(node));
 }
 
 /*
-*   @function includeContentInName
+*   @function getNodeContents
 *
-*   @desc Checks the CSS display and hidden properties, and
-*         the aria-hidden property to see if the content
-*         should be included in the accessible name
-*        calculation.  Returns true if it should be
-*         included, otherwise false
+*   @desc Get text content from a node for a name or description
 *
 *   @param  {Object}   node     -  DOM node
 *   @param  {Object}   forElem  -  DOM node the name is being computed for
@@ -536,7 +532,7 @@ function getNodeContents (node, forElem, alwaysInclude=false) {
               contents = getEmbeddedControlValue(node);
             }
             else {
-              if (node.hasChildNodes()) {
+              if (!isAriaHiddenTrue(node) && node.hasChildNodes()) {
                 let children = Array.from(node.childNodes);
                 children.forEach( child => {
                   [nc, nInclAlt, nInclAriaLabel] = getNodeContents(child, forElem);