+++ release-notes.md
@@ -5,6 +5,8 @@ Version 2.0.6 Overview
 * Updated documentation on the SEARCH element creating a search landmark
 * Added "isInClosedDetails" to visibility calculation
 * Updated accessible name calculation to include visibility for :before and :after content
+* Support inert attribute
+* Fixed support for aria-hidden=true attribute in accessible name calculation
 
 Version 2.0.5 Overview
 =============================